// frontend/src/app/utils/type.ts
export interface AddEthereumChainParameter {
  chainId: string;
  chainName: string;
  nativeCurrency: {
    name: string;
    symbol: string;
    decimals: number;
  };
  rpcUrls: string[];
  blockExplorerUrls?: string[];
}

export interface ProviderRpcError extends Error {
  code: number;
  data?: unknown;
}